Create test utils, clean up
