+++ typescript/sdk/test/core/message/interchain-gas.test.ts
@@ -2,140 +2,38 @@ import { formatMessage } from '@abacus-network/utils/dist/src/utils';
 import { expect } from 'chai';
 import { BigNumber, ethers, FixedNumber } from 'ethers';
 
-import { NameOrDomain } from '../../../src/types';
 import { AbacusCore, InterchainGasPayingMessage } from '../../../src/core';
-
-export const addresses = {
-  test1: {
-    upgradeBeaconController: '0x0000000000000000000000000000000000000000',
-    xAppConnectionManager: '0x0000000000000000000000000000000000000000',
-    validatorManager: '0x0000000000000000000000000000000000000000',
-    interchainGasPaymaster: '0x0000000000000000000000000000000000000000',
-    outbox: {
-      proxy: '0x0000000000000000000000000000000000000000',
-      implementation: '0x0000000000000000000000000000000000000000',
-      beacon: '0x0000000000000000000000000000000000000000',
-    },
-    inboxes: {
-      test2: {
-        proxy: '0x0000000000000000000000000000000000000000',
-        implementation: '0x0000000000000000000000000000000000000000',
-        beacon: '0x0000000000000000000000000000000000000000',
-      },
-    },
-  },
-  test2: {
-    upgradeBeaconController: '0x0000000000000000000000000000000000000000',
-    xAppConnectionManager: '0x0000000000000000000000000000000000000000',
-    validatorManager: '0x0000000000000000000000000000000000000000',
-    interchainGasPaymaster: '0x0000000000000000000000000000000000000000',
-    outbox: {
-      proxy: '0x0000000000000000000000000000000000000000',
-      implementation: '0x0000000000000000000000000000000000000000',
-      beacon: '0x0000000000000000000000000000000000000000',
-    },
-    inboxes: {
-      test1: {
-        proxy: '0x0000000000000000000000000000000000000000',
-        implementation: '0x0000000000000000000000000000000000000000',
-        beacon: '0x0000000000000000000000000000000000000000',
-      },
-    },
-  }
-}
-
-const MOCK_NETWORK = {
-  name: 'MockNetwork',
-  chainId: 1337,
-};
-
-class MockProvider extends ethers.providers.BaseProvider {
-
-  private methodResolveValues: { [key: string]: any };
-
-  constructor() {
-    super(MOCK_NETWORK);
-
-    this.methodResolveValues = {};
-  }
-
-  // Required to be implemented or the BaseProvider throws
-  async detectNetwork() {
-    return Promise.resolve(MOCK_NETWORK);
-  }
-  
-  perform(method: string, params: any): Promise<any> {
-    const value = this.methodResolveValues[method]
-    if (value) {
-      return Promise.resolve(value);
-    }
-
-    // switch (method) {
-    //   case 'getGasPrice':
-    //     return Promise.resolve(
-    //       BigNumber.from(12345)
-    //     );
-    //   case 'estimateGas':
-    //     return Promise.resolve(
-    //       BigNumber.from(21000)
-    //     );
-    // }
-    
-    return super.perform(method, params);
-  }
-
-  setMethodResolveValue(method: string, value: any) {
-    this.methodResolveValues[method] = value;
-  }
-
-  clearMethodResolveValues() {
-    this.methodResolveValues = {};
-  }
-}
-
-class TestTokenPriceGetter {
-  tokenPrices: { [domain: number]: FixedNumber }
-
-  constructor() {
-    this.tokenPrices = {};
-  }
-
-  getNativeTokenUsdPrice(domain: NameOrDomain): Promise<FixedNumber> {
-    const price = this.tokenPrices[domain as number];
-    if (price) {
-      return Promise.resolve(price);
-    }
-    throw Error(`No price for domain ${domain}`);
-  }
-}
+import { MockProvider, MockTokenPriceGetter, testAddresses } from '../../utils';
 
 describe('InterchainGasPayingMessage', () => {
+  const sourceDomain = 1;
+  const destinationDomain = 2;
 
   const testSerializedMessage = formatMessage(
-    1,
+    sourceDomain,
     ethers.constants.AddressZero,
     0,
-    2,
+    destinationDomain,
     ethers.constants.AddressZero,
     '0x12345678',
   );
 
   let core: AbacusCore;
   let provider: MockProvider;
-  let tokenPriceGetter: TestTokenPriceGetter;
+  let tokenPriceGetter: MockTokenPriceGetter;
   let testMessage: InterchainGasPayingMessage;
 
   before(() => {
-    core = new AbacusCore(addresses);
+    core = new AbacusCore(testAddresses);
     provider = new MockProvider();
     core.registerProvider('test1', provider);
     core.registerProvider('test2', provider);
 
-    tokenPriceGetter = new TestTokenPriceGetter();
+    tokenPriceGetter = new MockTokenPriceGetter();
     // Source domain token
-    tokenPriceGetter.tokenPrices[1] = FixedNumber.from(10);
+    tokenPriceGetter.setTokenPrice(sourceDomain, 10);
     // Destination domain token
-    tokenPriceGetter.tokenPrices[2] = FixedNumber.from(5);
+    tokenPriceGetter.setTokenPrice(destinationDomain, 5);
   });
 
   beforeEach(() => {
